- Adjusted further Visual C++ options
git-svn-id: svn+ssh://svn.morknet.de/usr/local/subversion/blz3/trunk@1000 c0f852eb-cc16-0410-a65c-258990c0cc53

+++ src/smart_gui/CoolMenu.h
@@ -111,7 +111,7 @@ protected:
 
 		if (lookup)
 		{
-			long int liVal = (long int)val;
+			long long liVal = (long long)val;
 
 			result = (int)(liVal & 0xffffffff);
 		}
